websocket device api was moved into separate namespace
