+++ devicehive/device/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+# vim:set et tabstop=4 shiftwidth=4 nu nowrap fileencoding=utf-8:
+
+__all__ = ['ws']
+
